Create module files for alias feature
All files with functions that for a feature will be structured this way
