+++ executor.c
@@ -1,16 +1,15 @@
-#include "simple_shell.h"
-#include "stdsh.h"
-#include "command.h"
-#include "utils.h"
-#include "types.h"
-
 #include <libgen.h>
+#include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <limits.h>
 #include <sys/wait.h>
 
+#include "alias.h"
+#include "command.h"
+#include "simple_shell.h"
+#include "stdsh.h"
+#include "utils.h"
